requirements.txt
altgraph==0.17.4
babel==2.17.0
charset-normalizer==3.4.3
macholib==1.16.3
packaging==25.0
pillow==11.3.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pytz==2025.2
reportlab==4.4.4
setuptools==80.9.0
tk==0.1.0
tkcalendar==1.6.1
